[refactor] 컴포넌트 구조 개선 - Projects, Skills, About 컴포넌트를 서버 컴포넌트와 클라이언트 컴포넌트로 분리 - 상호작용이 없는 Footer 컴포넌트를 서버 컴포넌트로 전환 - page.tsx에 있던 데이터 로직을 해당 컴포넌트(project, skill)로 이동
