+++ src/components/layout/Footer.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 export function Footer() {
   return (
     <footer
