+++ src/components/sections/Projects.tsx
@@ -1,24 +1,19 @@
-"use client";
-
 import { AnimatedSection } from "../common/AnimatedSection";
-import { Project } from "@/types";
-import { useState } from "react";
-import { ProjectCard } from "../ui/ProjectCard";
-import { ProjectModal } from "../ui/ProjectModal";
+import { ProjectList } from "./ProjectList";
+import prisma from "@/lib/prisma";
 
-export function Projects({ projects }: { projects: Project[] }) {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [selectedProject, setSelectedProject] = useState<Project | null>(null);
+async function getProjects() {
+  const projects = await prisma.project.findMany({
+    include: {
+      info: true,
+      tasks: true,
+    },
+  });
+  return projects;
+}
 
-  const handleOpenModal = (project: Project) => {
-    setSelectedProject(project);
-    setIsModalOpen(true);
-  };
-
-  const handleCloseModal = () => {
-    setIsModalOpen(false);
-    setSelectedProject(null);
-  };
+export async function Projects() {
+  const projects = await getProjects();
 
   return (
     <section
@@ -30,24 +25,8 @@ export function Projects({ projects }: { projects: Project[] }) {
           <h2 className={`section-header text-3xl md:text-4xl `}>Projects</h2>
         </AnimatedSection>
 
-        <div className={`grid gap-8 md:grid-cols-2 lg:grid-cols-3`}>
-          {projects.map((project) => (
-            <AnimatedSection
-              key={project.id}
-              animation="slide-up"
-              delay={parseFloat(project.id) * 0.1}
-            >
-              <ProjectCard
-                openDetailModal={() => handleOpenModal(project)}
-                project={project}
-              />
-            </AnimatedSection>
-          ))}
-        </div>
+        <ProjectList projects={projects} />
       </div>
-      {isModalOpen && selectedProject && (
-        <ProjectModal project={selectedProject} onClose={handleCloseModal} />
-      )}
     </section>
   );
 }
