+++ src/hooks/useMediaQuery.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useState } from "react";
 
 export function useMediaQuery(query: string) {
